src/requirements.txt
-i https://pypi.doubanio.com/simple/
Flask==1.0.3
gunicorn==19.7.1
numpy==1.16.3
pandas==0.24.2
python-dateutil==2.8.0
python-logstash-async==1.6.4
pytz==2019.1
requests==2.18.1
six==1.13.0
SQLAlchemy==1.3.17
pymysql==0.9.3
pyecharts==1.7.1
jsonpath==0.82
openpyxl==3.0.3
Faker==0.7.11
fake2db==0.5.4
py_eureka_client==0.7.4
scipy
aurora-common-utils
flask_sqlalchemy
xlwt
xlrd==1.0.0
lxml==4.6.2
html5lib==1.0.1
beautifulsoup4==4.6.3